+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is (loosely) based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased] - 
+
+## [Unreleased]
+### Fixed
+- fix catalog file name check https://github.com/stac-utils/stac-check/pull/83
 
 ## [v1.2.0] - 2022-04-26
 ### Added